### 2026-01-13 #9772 @editorialbot
[REVIEW]: Cont2SAS: A python package for calculating SAS parameters from continuum nanostructures
**Submitting author:** @majumdararn (<a href="http://orcid.org/0000-0003-4049-4060">Arnab Majumdar</a>)
**Repository:** https://github.com/hereon-GEMS/Cont2SAS
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.2
