### 2026-01-22 #9772 @kremeyer
Do I need to regenerate my checklist to have it reflect the pre-2026 scope?
