+++ models/sld_grow/sld_grow_gen.py
@@ -80,7 +80,7 @@ qclean_sld=sld_out
 ### scatt cal ###
 num_cat=101 # also check with 3
 method_cat='extend'
-sassena_exe= '/home/amajumda/Documents/Softwares/sassena/compile/sassena'
+sassena_exe= '/home/amajumda/Dokumente/Software/source_code/sassena/build-rel-cpu/bin/sassena'
 mpi_procs=1
 num_threads=1
 sig_file='signal.h5'
