### 2026-01-27 2f26d25
default time values are ini anf final
