### 2026-01-27 8ba2787
ini ver of ci script
