+++ .gitlab-ci.yml
@@ -1,39 +1,72 @@
-# .gitlab-ci.yml
+name: CI
 
-stages:
-  - build
-  - test
+on:
+  push:
+    branches:
+      - main
+      - develop 
+  pull_request:
 
-download_sassena:
-  stage: build
-  image: ubuntu:latest
-  before_script:
-    - apt-get update && apt-get install -y curl
-  script: 
-    - mkdir -p ./Sassena
-    - curl -L https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage -o Sassena/Sassena.AppImage
-    - chmod +x ./Sassena/Sassena.AppImage
-  artifacts:
-    paths:
-      - "Sassena"
+concurrency:
+  group: ci-${{ github.ref }}
+  cancel-in-progress: true
 
-run_tests:
-  stage: test
-  image: ubuntu:latest
-  dependencies:
-    - download_sassena
-  before_script:
-    - apt-get update && apt-get install -y mpi-default-dev wget ghostscript
-    - export APPIMAGE_EXTRACT_AND_RUN=1
-  script:
-    - wget -O Miniforge3.sh "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
-    - bash Miniforge3.sh -b -p "${HOME}/conda"
-    - source "${HOME}/conda/etc/profile.d/conda.sh"
-    - conda activate
-    - cd /builds/arnab.majumdar/continuum-to-scattering/
-    - export C2S_HOME=$PWD
-    - conda env create -f environment.yml
-    - conda activate Cont2Sas
-    - conda install pytest
-    - pytest -v -rsx
+jobs:
+  test:
+    name: Build & Test
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
 
+    env:
+      APPIMAGE_EXTRACT_AND_RUN: 1
+      C2S_HOME: ${{ github.workspace }}
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      # -----------------------
+      # System dependencies
+      # -----------------------
+      - name: Install system packages
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            curl \
+            wget \
+            ghostscript \
+            mpi-default-dev
+
+      # -----------------------
+      # Download Sassena
+      # -----------------------
+      - name: Download Sassena AppImage
+        run: |
+          mkdir -p Sassena
+          curl -L \
+            https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage \
+            -o Sassena/Sassena.AppImage
+          chmod +x Sassena/Sassena.AppImage
+
+      # -----------------------
+      # Miniforge + Conda cache
+      # -----------------------
+      - name: Setup Miniforge
+        uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniforge-variant: Miniforge3
+          miniforge-version: latest
+          auto-activate-base: false
+          activate-environment: Cont2Sas
+          environment-file: environment.yml
+          use-mamba: true
+          cache-environment: true
+
+      # -----------------------
+      # Tests
+      # -----------------------
+      - name: Run tests
+        shell: bash -l {0}
+        run: |
+          conda install -y pytest
+          pytest -v -rsx
