### 2026-01-27 #9772 @jgostick
So it seems that @ebknudsen was able to create issues on the repo as it's hosted now. Perhaps @kremeyer could try logging in with github credentials too?  I tried it using my github for auth, and it seemed to work.

I think if we can raise issues on the repo where it's hosted now, then that meets the joss criteria. I haven't tried to make a PR but I was able to start an issue.

So let's use the helmholz gitlab for joss related issues.
