### 2026-01-27 5a930bb
ini ver of ci script
