### 2026-01-27 #9772 @kremeyer
Yes, issues work for me. I interpreted the phrase "users can browse the source code files, open issues, and **propose code changes**" from the [JOSS submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) as a requirement for PRs. Since that seems not to be the case, I'm happy to explain the bugs I found in an issue.
