### 2026-01-27 dcb11b1
removed gitlab ci after migration to github
