### 2026-01-28 423e276
reviewer comment: figure fonts too small
