### 2026-01-28 9e47fb7
edited paper fig2
