+++ joss_submission/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 The simulation of material structure at the nanometer length scale can be performed using atomistic simulations and continuum simulations. Continuum simulations have the advantage over atomistic ones that they can simulate bigger volumes for a larger time. However, continuum simulations are less accurate than the atomistic simulations. `Cont2SAS` is created to check the accuracy of continuum simulations by validating them against SAS experiments, such as Small Angle Neutron Scattering (SANS) and Small Angle X-ray Scattering (SAXS). The validation is performed by comparing SAS parameters calculated using `Cont2SAS` with measured ones. A validated simulation can also be used to retrieve nanostructure from SAS data because a direct retrieval of nanstructure from SAS data is not possible [@billinge2007problem]. Alternatively to simulating continuum nanostructures based on physics-based equations, they can also be simulated based on the user's knowledge of the sample to retrieve nanostructures from SAS data.
 
-![Workflow of SAS pattern calculation: [left] Mesh generation, [middle] SLD assignment, [right] SAS pattern calculation. \label{fig:sas_workflow}](figures/workflow.png)
+![Workflow of SAS pattern calculation: [left] Mesh generation, [middle] SLD assignment, [right] numerical calculation of SAS pattern. The numerical calculation matches well with the known analytical formula for spherical nanoparticles [@guinier1955small]. \label{fig:sas_workflow}](figures/workflow.png)
 
 Both physics- and knowledge-based simulated structures are expected to output either Scattering Length Density (SLD) ($\beta$) values or a set of variables (e.g. local molar density ($\rho_{\text{m}}$) and composition ($\chi$)) from which SLD values can be calculated. `Cont2SAS` creates a mesh, assigns SLD based on the simulated values, and calculates SAS pattern from them, as shown in \autoref{fig:sas_workflow}.
 
