+++ joss_submission/paper.bib
@@ -79,3 +79,9 @@ doi = {10.3233/JNR-190116}
   year  = {2017},
   howpublished   = "\url{https://github.com/camm/sassena}",
 }
+@book{guinier1955small,
+  title={Small-angle Scattering of X-rays},
+  author={Guinier, Andr{\'e} and Fournet, G{\'e}rard and Walker, Christopher B and Yudowitch, Kenneth L},
+  year={1955},
+  publisher={Wiley New York}
+}
