### 2026-01-28 4359ece
added ref for ana formula of sph
