+++ joss_submission/paper.bib
@@ -159,3 +159,9 @@ doi = {10.3233/JNR-190116}
   year  = {2023},
   howpublished   = "\url{https://github.com/Z-Laboratory/LiquidLib/tree/master}",
 }
+@book{guinier1955small,
+  title={Small-angle Scattering of X-rays},
+  author={Guinier, Andr{\'e} and Fournet, G{\'e}rard and Walker, Christopher B and Yudowitch, Kenneth L},
+  year={1955},
+  publisher={Wiley New York}
+}
