### 2026-01-28 c312a7b
changed fig1 caption, consistent with arnab branch
