### 2026-01-28 ef732b2
copied paper md and bib from dev branch
