+++ joss_submission/paper.bib
@@ -1,3 +1,25 @@
+@article{chen2012characterizing,
+  title={Characterizing size and porosity of hollow nanoparticles: SAXS, SANS, TEM, DLS, and adsorption isotherms compared},
+  author={Chen, Zhi Hong and Kim, Chanhoi and Zeng, Xiang-bing and Hwang, Sun Hye and Jang, Jyongsik and Ungar, Goran},
+  journal={Langmuir},
+  volume={28},
+  number={43},
+  pages={15350--15361},
+  year={2012},
+  publisher={ACS Publications},
+  doi={10.1021/la302236u}
+}
+@article{hollamby2013practical,
+  title={Practical applications of small-angle neutron scattering},
+  author={Hollamby, Martin J.},
+  journal={Phys. Chem. Chem. Phys.},
+  volume={15},
+  issue={26},
+  pages={10566--10579},
+  year={2013},
+  publisher={The Royal Society of Chemistry},
+  doi={10.1039/C3CP50293G}
+}
 @article{billinge2007problem,
   title={The problem with determining atomic structure at the nanoscale},
   author={Billinge, Simon JL and Levin, Igor},
@@ -6,7 +28,15 @@
   number={5824},
   pages={561--565},
   year={2007},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1135080}
+}
+@book{ghavanloo2019computational,
+  title={Computational continuum mechanics of nanoscopic structures},
+  author={Ghavanloo, Esmaeal and Rafii-Tabar, Hashem and Fazelzadeh, Seyed Ahmad},
+  year={2019},
+  publisher={Springer},
+  doi={10.1007/978-3-030-11650-7}
 }
 @article{aslan2019high,
 title ={High-pressure cell for in situ neutron studies of hydrogen storage materials},
@@ -78,6 +108,56 @@ doi = {10.3233/JNR-190116}
   title = {Sassena},
   year  = {2017},
   howpublished   = "\url{https://github.com/camm/sassena}",
+}
+@article{rog2003nmoldyn,
+  title={nMoldyn: a program package for a neutron scattering oriented analysis of molecular dynamics simulations},
+  author={R{\'o}g, Tomasz and Murzyn, Krzysztof and Hinsen, Konrad and Kneller, Gerald R},
+  journal={Journal of computational chemistry},
+  volume={24},
+  number={5},
+  pages={657--667},
+  year={2003},
+  publisher={Wiley Online Library},
+  doi={10.1002/jcc.10243}
+}
+@misc{nmoldyn2023hinsen,
+  author = {Hinsen, Konrad},
+  title = {nMOLDYN3--Version 3.0.12},
+  year  = {2023},
+  howpublished   = "\url{https://github.com/khinsen/nMOLDYN3/}",
+}
+@article{goret2017mdanse,
+  title={MDANSE: An interactive analysis environment for molecular dynamics simulations},
+  author={Goret, G and Aoun, B and Pellegrini, Eric},
+  journal={Journal of chemical information and modeling},
+  volume={57},
+  number={1},
+  pages={1--5},
+  year={2017},
+  publisher={ACS Publications},
+  doi={10.1021/acs.jcim.6b00571}
+}
+@misc{mdanse2023isis,
+  author = {UKRI STFC ISIS Neutron and Muon Facility},
+  title = {MDANSE--Version 1.5.2},
+  year  = {2023},
+  howpublished   = "\url{https://github.com/ISISNeutronMuon/MDANSE/}",
+}
+@article{walter2018liquidlib,
[…]
