+++ joss_submission/paper.md
@@ -29,31 +29,35 @@ bibliography: paper.bib
 
 # Summary
 
-`Cont2SAS` facilitates the calculation of Small Angle Scattering (SAS) parameters from simulated Continuum (Cont) nanostructures. `Cont2SAS` is built on the existing software solution `Sassena` -- known for calculating scattering patterns from simulated atomic structures [@lindner2012sassena; @lindner2012towards; @sassena2017lindner; @sassena2023majumdar; @majumdar2024computation]. `Cont2SAS` can calculate SAS patterns and the effective scattering cross-section $(\sigma_{\text{eff}})$. SAS patterns contain a SAS intensity $(I)$ vs. scattering vector magnitude $(Q)$. The $\sigma_{\text{eff}}$ is the count rate of scattered radiation per incident unit flux. The time evolution of $\sigma_{\text{eff}}$ is calculated from SAS patterns at different time steps. Through the comparison of calculated and measured SAS parameters, simulations and SAS experiments can be used complementarily for different purposes, such as validating simulations, tuning simulation parameters, and analyzing SAS data obtained from experiments [@dorrell2020combined; @reich2022comparison; @majumdar2024computation].
+`Cont2SAS` is a software tool built around the existing software solution `Sassena` -- known for calculating scattering patterns from simulated atomic structures. The goal of `Cont2SAS` is to provide a software similar to `Sassena`, but for calculating Small Angle Scattering (SAS) data from simulated Continuum (Cont) nanostructures. `Cont2SAS` can calculate SAS patterns, i.e. SAS intensity $(I)$ vs. scattering vector magnitude $(Q)$ using a novel numerical method. `Cont2SAS` can also calculate effective cross-section $(\sigma_{\text{eff}})$, i.e. the count rate of scattered radiation per incident unit flux. $\sigma_{\text{eff}}$ is calculated by integrating instrument-agnostic SAS patterns taking instrument geometry into account. `Cont2SAS` can be used for different purposes, such as validating simulations, tuning simulation parameters, and analyzing SAS data.
 
 # Statement of need
 
-The simulation of material structure at the nanometer length scale can be performed using atomistic simulations and continuum simulations. Continuum simulations have the advantage over atomistic ones that they can simulate bigger volumes for a larger time. However, continuum simulations are less accurate than the atomistic simulations. `Cont2SAS` is created to check the accuracy of continuum simulations by validating them against SAS experiments, such as Small Angle Neutron Scattering (SANS) and Small Angle X-ray Scattering (SAXS). The validation is performed by comparing SAS parameters calculated using `Cont2SAS` with measured ones. A validated simulation can also be used to retrieve nanostructure from SAS data because a direct retrieval of nanstructure from SAS data is not possible [@billinge2007problem]. Alternatively to simulating continuum nanostructures based on physics-based equations, they can also be simulated based on the user's knowledge of the sample to retrieve nanostructures from SAS data.
+Small Angle Scattering (SAS) experiments with neutrons (SANS) and X-rays (SAXS) are useful techniques for probing material nanostructures [@chen2012characterizing]. With the Time Resolved (TR) variant of SAS, the time evolution of nanostructures can also be probed [@hollamby2013practical]. However, a direct retrieval of nanostructure from SAS data is not possible due to the so-called phase problem [@billinge2007problem]. Therefore, it is fruitful to combine SAS data with simulations to study nanostructures [@majumdar2024computation]. This approach is also useful for validating theories underlying simulations [@reich2022comparison].
+
[…]
