### 2026-01-28 8976034
fig gen scripts for joss copied from M-Arnab
