### 2026-01-28 1b7350d
added ball plot script for joss
