+++ models/ball/ball_plot_joss.py
@@ -1,0 +1,322 @@
+# Copyright (C) 2025  Helmholtz-Zentrum-Hereon
+
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, see
+# <https://www.gnu.org/licenses/>.
+
+"""
+This plots necessary figures for ball model
+Plots are saved in figure folder
+
+Author: Arnab Majumdar
+Date: 24.06.2025
+"""
+import sys
+import os
+import warnings
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.patches import Rectangle
+import h5py
+
+
+# find current dir and and ..
+lib_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '../..'))
+# add .. in path
+sys.path.append(lib_dir)
+# lib imports (if any)
+
+# ignore warnings
+warnings.filterwarnings("ignore")
+
+# analytical SAS function
+def J1(x):
+    """
+    function desc:
+    spherical bessel function
+    """
+    if x==0:
+        y = 0
+    else:
+        y = (np.sin(x)-x*np.cos(x))/x**2
+    return y
+
+def ball (qmax,qmin,Npts,scale,bg,sld,sld_sol,rad):
+    """
+    function desc:
+    analytical SAS pattern of sphere
+    """
+    vol=(4/3)*np.pi*rad**3
+    # SLD unit 10^-5 \AA^-2
+    del_rho=sld-sld_sol
+    q_arr=np.linspace(qmin,qmax,Npts)
+    FormFactor=np.zeros(len(q_arr))
+    for q_idx, q_cur in enumerate(q_arr):
+        if q_cur==0:
+            # Form factor unit 10^-5 \AA
+            FormFactor[q_idx]=vol*del_rho
+        else:
+            # Form factor unit 10^-5 \AA
+            FormFactor[q_idx]=3*vol*del_rho*J1(q_cur*rad)/(q_cur*rad)
+    Iq_arr = (scale)*np.abs(FormFactor)**2+bg
+    return Iq_arr, q_arr
+
+##########################
+# read input from xml file
+##########################
+
+### file locations ###
+# xml location
+xml_folder='./xml/'
+# script dir and working dir
+script_dir = "src"  # Full path of the script
+working_dir = "."  # Directory to run the script in
+
+############
+# Input data
+############
+
+### struct gen ###
+# box side lengths (float values)
+length_a= 40.
+length_b=40.
+length_c=40.
+# number of cells in each direction (int values)
+nx= 40
+ny= 40
+nz= 40
+# element details
+el_type='lagrangian'
+el_order=1
+# calculate mid point of structure (simulation box)
+mid_point=np.array([length_a/2, length_b/2, length_c/2])
+
+### sim gen ###
+# model name
+sim_model= 'ball'
+# simulation parameters
+dt= 1.
+t_end= 0.
+n_ensem=1
+# calculate time array
+t_arr=np.arange(0,t_end+dt, dt)
+
+### model param ###
+# model params
+ball_rad=10
+ball_sld=2
+qclean_sld=0
+# dir name for model param
[…]
