### 2026-01-28 46c52d0
fs plot inp params consistent with M-Arnab
