### 2026-01-28 2421b41
phase field plot consistent with M-Arnab
