### 2026-01-28 3e15c11
edits in paper after auth 3 comment
