+++ joss_submission/paper.md
@@ -47,7 +47,7 @@ From atomistic simulations, different software such as `nMoldyn` [@rog2003nmoldy
 
 `Cont2SAS` calculates SAS patterns taking simulated nanostructures as an input. The simulated structure must provide either Scattering Length Density (SLD) ($\beta$) values or a set of variables from which SLD values can be calculated. The simulated input is processed to a data taylor-made for `Sassena`. `Sassena` calculates the SAS intensity $(I)$ as a function of scattering vector magnitude $(Q)$, i.e. SAS pattern, for different time steps (see \autoref{fig:sas_workflow}). 
 
-![Workflow of effective cross-section $(\sigma_{\text{eff}})$ calculation: [left] SAS patterns at different timesteps ranging from 0 to 10 seconds and [right] $\sigma_{\text{eff}}$ calculated from SAS patterns. With time, the SLD of a spherical nanoparticle increases linearly. The calculated $\sigma_{\text{eff}}$ values are proportional to the square of SLD difference between the particle and its environment, shown as contrast. \label{fig:sig_eff_workflow}](figures/sig_eff.png)
+![Workflow of effective cross-section $(\sigma_{\text{eff}})$ calculation: [left] SAS patterns at different timesteps ranging from 0 to 10 seconds and [right] $\sigma_{\text{eff}}$ calculated from SAS patterns. With time, the SLD of a spherical nanoparticle increases linearly. The calculated $\sigma_{\text{eff}}$ values are proportional to the ($\Delta$SLD)$^2$, i.e. square of SLD difference between the particle and its environment. \label{fig:sig_eff_workflow}](figures/sig_eff.png)
 
 `Cont2SAS` can also calculate the time evolution of effective cross-section $(\sigma_{\text{eff}})$, which is defined as the count rate per incident unit flux. This feature was not available in the in-house code [@dorrell2020combined] or `Sassena` [@sassena2023majumdar]. The time evolution of count rate is useful when the chemical composition changes over time instead of the nanostructure, e.g. while storing hydrogen in ball-milled powder sample [@aslan2019high]. \autoref{fig:sig_eff_workflow} demonstrates the workflow of such calculation from a series of SAS patterns. The calculated $\sigma_{\text{eff}}$ must be multiplied by an empirical factor before comparing with measured count rate.   
 
@@ -62,6 +62,6 @@ Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) - project numb
 
 # Declaration
 
-During the preparation of this work, the authors used Chat-GPT in order to improve the quality of code and article. After using this tool, the authors reviewed and edited the content as needed and take full responsibility for the content of the published code and article.
+During the preparation of this work, the authors used ChatGPT in order to improve the quality of code and article. After using this tool, the authors reviewed and edited the content as needed and take full responsibility for the content of the published code and article.
 
 # References
