+++ joss_submission/paper.md
@@ -64,5 +64,4 @@ Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) - project numb
 
 During the preparation of this work, the authors used ChatGPT in order to improve the quality of code and article. After using this tool, the authors reviewed and edited the content as needed and take full responsibility for the content of the published code and article.
 
-
 # References
