### 2026-01-29 9b30aaf
fixed typo in contr.md
