### 2026-01-29 eed0213
deleted comments in contr.md
