### 2026-01-29 hereon-GEMS/Cont2SAS#3 @majumdararn
JOSS submission: Figures should be improved
This issue carried forward while migration from GitLab to GitHub.
This issue was raised while reviewing the code in GitLab (see [link](https://codebase.helmholtz.cloud/arnab.majumdar/continuum-to-scattering/-/issues/2)).

Reviewer @ebknudsen comment:

> Axis on figures 1 and 2 are almost illegible - please use a bigger font, or add some more text in the captions. As is, they're almost illegible, without zooming in a lot.
>
>Also you should add some more detail as to what they actually show. E.g. figure 1, right: There's very good agreement between an analytical and a numerical calculation but what are they?
>
>I'd suggest removing figure 1, left. It has no axis and hence do not convey really convey any useful information. From just reading the text there's no reference to a mesh being created, much less what it's used for.

Reviewer @kremeyer comment:

> I had some issues with the Figures as well and I'm going to add some points here. I fully agree with @erik's points.
>
> In Figure 2 (left) you use different colors for presumably different time-steps? This needs to be clarified. And on the right panel you are plotting some form of fit that is not described in the caption or manuscript.I had some issues with the Figures as well and I'm going to add some points here.
