### 2026-01-29 #9772 @majumdararn
Dear all,

Thank you very much for all your comments and raised issues in [helmholtz codebase](https://codebase.helmholtz.cloud/arnab.majumdar/continuum-to-scattering).

I have migrated the code from [helmholtz codebase](https://codebase.helmholtz.cloud/arnab.majumdar/continuum-to-scattering) to [helmholtz hereon GitHub organization](https://github.com/hereon-GEMS). The new URL is [https://github.com/hereon-GEMS/Cont2SAS](https://github.com/hereon-GEMS/Cont2SAS).

I have created a tag v1.0.0, which includes the following changes:
1. Removed .gitlab-ci.yml, included .github/workflows/ci.yml.
2. Edited the contributors.md and changed the first code contribution section to cater for GitHub workflow.
3. Replaced all occurrences of the old repository URL with the new repository URL.
4. Moved all the issues raised by reviewers to the new location.

Reviewer comments:
1. I have already implemented the necessary changes mentioned in [issue 2](https://codebase.helmholtz.cloud/arnab.majumdar/continuum-to-scattering/-/issues/2). I will make a comment regarding the changes made in the [corresponding github issue](https://github.com/hereon-GEMS/Cont2SAS/issues/3#issue-3870170999).
2. Other issues will be addressed as soon as possible. I will implement the changes in the develop branch, merge them to main, and create a new tag.

With regards,
@majumdararn
