### 2026-02-02 a899e5b
brought matplotlib ver down, consistent with env.yml
