### 2026-02-02 d911daa
made all versions consistent with env.yml, tested
