+++ requirements.txt
@@ -3,9 +3,9 @@ imageio==2.33.1
 matplotlib==3.8.2
 mdtraj==1.9.9
 netCDF4==1.7.1
-numpy==1.24.4
+numpy==1.26.4
 pdf2image==1.17.0
-PyMuPDF==1.24.11
+PyMuPDF==1.26.1
 pylint==3.3.8
 pytest==7.4.4
-scipy==1.10.1
+scipy==1.11.4
