### 2026-02-02 912b6dc
updated sassena installation instructions
