### 2026-02-03 b0f1d88
remove plot comp, add data compare
