### 2026-02-03 c22c435
removed plot compare, added data compare
