+++ tests/test_sld_grow.py
@@ -27,8 +27,24 @@ Date      : 06.08.2025
 import shutil
 import os
 import subprocess
-from matplotlib.testing.compare import compare_images as comp_img
+import pytest
+# from matplotlib.testing.compare import compare_images as comp_img
+import h5py
 
+# non-test functions
+def read_sig_eff_h5(sig_eff_data_h5):
+    """
+    Function for 
+    1. Reading sig_eff data (sig_eff_data_h5)
+    2. normalize using vol_norm
+    """
+    sig_eff_data=h5py.File(sig_eff_data_h5,'r')
+    sig_eff=sig_eff_data['sig_eff'][:]
+    t=sig_eff_data['t'][:]
+    sig_eff_data.close()
+    return t, sig_eff
+
+# test functions
 def test_sld_grow_gen():
     """
     Test script for $C2S_HOME/models/sld_grow/sld_grow_gen.py:
@@ -67,15 +83,34 @@ def test_check_sld_grow_plot():
     # Ensure plotted file exist
     assert if_exist==1
 
-def test_compare_sld_grow_plot():
+# def test_compare_sld_grow_plot():
+#     """
+#     Test script for $C2S_HOME/models/sld_grow/sld_grow_plot.py:
+#     1. Compare plotted IvsQ figure with gold standard
+#     """
+#     gold_img='figure/gold/sig_eff_fit_sld_grow.pdf'
+#     plt_img='figure/sld_grow/sig_eff_fit_sld_grow.pdf'
+#     # Check expected output
+#     assert comp_img(plt_img, gold_img, tol=2) is None
+
+def test_compare_sld_grow_data():
     """
-    Test script for $C2S_HOME/models/sld_grow/sld_grow_plot.py:
-    1. Compare plotted IvsQ figure with gold standard
+    Test script for $C2S_HOME/models/sld_grow/sld_grow_gen.py:
+    1. Compare numerical IvsQ data with gold standard
     """
-    gold_img='figure/gold/sig_eff_fit_sld_grow.pdf'
-    plt_img='figure/sld_grow/sig_eff_fit_sld_grow.pdf'
-    # Check expected output
-    assert comp_img(plt_img, gold_img, tol=2) is None
+    # folder structure of gen data
+    mesh_dir=os.path.join('data', '40p0_40p0_40p0_40_40_40_lagrangian_1')
+    sim_dir=os.path.join(mesh_dir, 'simulation/sld_grow_tend_10p0_dt_1p0_ensem_1')
+    model_dir=os.path.join(sim_dir, 'rad_10_sld_in_0_2_sld_in_end_5_sld_out_1_qclean_sld_1')
+    # t_dir=os.path.join(model_dir, 't000')
+    # data
+    gold_data='tests/gold/sld_grow_gold.h5'
+    t_gold, sig_eff_gold=read_sig_eff_h5(gold_data)
+    num_data=os.path.join(model_dir, 'sig_eff_cat_extend_101Q_0p0029_0p05_orien__10.h5')
+    t_num, sig_eff_num=read_sig_eff_h5(num_data)
+    # compare num value and gold value
+    assert t_num == pytest.approx(t_gold, abs=1e-6), "t values match"
+    assert sig_eff_num == pytest.approx(sig_eff_gold, abs=1e-6), "sig eff values match"
 
 def test_clean_up():
     """
