### 2026-02-03 1fe53d2
remove plot compare, add data compare
