+++ tests/test_phase_field.py
@@ -28,20 +28,38 @@ import shutil
 import os
 import subprocess
 from matplotlib.testing.compare import compare_images as comp_img
+import pytest
+import h5py
 
-def test_phase_field_gen():
+# non-test functions
+def read_Iq_h5(Iq_data_h5, vol_norm):
     """
-    Test script for $C2S_HOME/models/phase_field/phase_field_gen.py:
-    1. Check whether phase_field_gen runs
+    Function for 
+    1. Reading Iq data (Iq_data_h5)
+    2. normalize using vol_norm
     """
-    gen_result = subprocess.run(
-        ["python", "models/phase_field/phase_field_gen.py"],
-          capture_output=True,
-            text=True,
-              check=True
-    )
-    # Ensure it runs without crashing
-    assert gen_result.returncode == 0, "phase_field model generates data"
+    Iq_data=h5py.File(Iq_data_h5,'r')
+    Iq=Iq_data['Iq'][:] # unit fm^2
+    Iq=Iq/vol_norm*10**2 # unit (fm^2 / \AA^3) * 10^2 = cm^-1
+    q=Iq_data['Q'][:]
+    Iq_data.close()
+    return q, Iq
+
+# test functions
+
+# def test_phase_field_gen():
+#     """
+#     Test script for $C2S_HOME/models/phase_field/phase_field_gen.py:
+#     1. Check whether phase_field_gen runs
+#     """
+#     gen_result = subprocess.run(
+#         ["python", "models/phase_field/phase_field_gen.py"],
+#           capture_output=True,
+#             text=True,
+#               check=True
+#     )
+#     # Ensure it runs without crashing
+#     assert gen_result.returncode == 0, "phase_field model generates data"
 
 def test_phase_field_plot():
     """
@@ -77,18 +95,47 @@ def test_compare_phase_field_plot():
     # Check expected output
     assert comp_img(plt_img, gold_img, tol=2) is None
 
-def test_clean_up():
+def test_compare_phase_field_data():
     """
-    Test script for clean up:
-    1. Removes created data and figure
-    2. Checks whether removed
+    Test script for $C2S_HOME/models/phase_field/phase_field_plot.py:
+    1. Compare numerical IvsQ data with gold standard
     """
-    data_dir='data/250p0_250p0_250p0_100_100_100_lagrangian_1'
-    fig_dir='figure/phase_field'
-    # remove data files
-    shutil.rmtree(data_dir)
-    # remove figure
-    shutil.rmtree(fig_dir)
-    # check if it is removed
-    assert os.path.isdir(data_dir)==0
-    assert os.path.isdir(fig_dir)==0
+    # folder structure of gen data
+    mesh_dir=os.path.join('data', '250p0_250p0_250p0_100_100_100_lagrangian_1')
+    sim_dir=os.path.join(mesh_dir, 'simulation/phase_field_tend_0p0_dt_1p0_ensem_1')
+    model_dir1=os.path.join(sim_dir, 'name_spinodal_fe_cr_time_0_qclean_sld_0p5410058695982137')
+    t_dir1=os.path.join(model_dir1, 't000')
+    model_dir2=os.path.join(sim_dir, 'name_spinodal_fe_cr_time_86400_qclean_sld_0p5408560149631366')
+    t_dir2=os.path.join(model_dir2, 't000')
+    # data 1
+    gold_data_1='tests/gold/phase_field_gold_1.h5'
+    q_gold_1, Iq_gold_1=read_Iq_h5(gold_data_1, 1)
[…]
