### 2026-02-04 2691f3a
uncomment gen and clean up tests
