### 2026-02-04 dbd5886
removed norm of Iq by volume
