### 2026-02-04 19a7b5e
added exist checks, clean code
