### 2026-02-04 75df15a
added checks in data compare, clean code
