+++ tests/test_phase_field.py
@@ -97,32 +97,62 @@ def test_compare_phase_field_plot():
 
 def test_compare_phase_field_data():
     """
-    Test script for $C2S_HOME/models/phase_field/phase_field_plot.py:
+    Test script for $C2S_HOME/models/phase_field/phase_field_gen.py:
     1. Compare numerical IvsQ data with gold standard
     """
     # folder structure of gen data
+    ## meshing dir
     mesh_dir=os.path.join('data', '250p0_250p0_250p0_100_100_100_lagrangian_1')
+    mesh_exist=os.path.isdir(mesh_dir)
+    assert mesh_exist==1, "meshing dir exist"
+    ## sim time = 0s
+    ## C2S creates model for each simulated time step
+    ## each simulated time step then have one C2S time step
+    ## this algo was in place because C2S only allows for linspace time arr
+    ### model dir 1
     sim_dir=os.path.join(mesh_dir, 'simulation/phase_field_tend_0p0_dt_1p0_ensem_1')
-    model_dir1=os.path.join(sim_dir, 'name_spinodal_fe_cr_time_0_qclean_sld_0p5410058695982137')
-    t_dir1=os.path.join(model_dir1, 't000')
-    model_dir2=os.path.join(sim_dir, 'name_spinodal_fe_cr_time_86400_qclean_sld_0p5408560149631366')
-    t_dir2=os.path.join(model_dir2, 't000')
-    # data 1
+    model_dir_1=os.path.join(sim_dir, 'name_spinodal_fe_cr_time_0_qclean_sld_0p5410058695982137')
+    model_exist_1=os.path.isdir(model_dir_1)
+    assert model_exist_1==1, "model dir exist, sim time 0"
+    ### time dir 1
+    t_dir_1=os.path.join(model_dir_1, 't000')
+    t_exist_1=os.path.isdir(t_dir_1)
+    assert t_exist_1==1, "time dir exist, sim time 0"
+    ### data
+    #### num data (read + check exist)
+    num_data_1=os.path.join(t_dir_1, 'Iq_cat_extend_501Q_0p025132741228718346_1p0_orien__200.h5')
+    data_exist_1=os.path.isfile(num_data_1)
+    assert data_exist_1==1, "hdf5 file exist, sim time 0"
+    q_num_1, Iq_num_1=read_Iq_h5(num_data_1, 1)
+    #### gold data
     gold_data_1='tests/gold/phase_field_gold_1.h5'
     q_gold_1, Iq_gold_1=read_Iq_h5(gold_data_1, 1)
-    num_data_1=os.path.join(t_dir1, 'Iq_cat_extend_501Q_0p025132741228718346_1p0_orien__200.h5')
-    q_num_1, Iq_num_1=read_Iq_h5(num_data_1, 1)
-    # compare num value and gold value
-    assert q_num_1 == pytest.approx(q_gold_1, abs=1e-6), "Q values match"
-    assert Iq_num_1 == pytest.approx(Iq_gold_1, abs=1e-6), "IQ values match"
-    # data 2
+    ## sim time = 86400s
+    ### model dir 2
+    ### sim dir is same
+    model_dir_2=os.path.join(sim_dir, 'name_spinodal_fe_cr_time_86400_qclean_sld_0p5408560149631366')
+    model_exist_2=os.path.isdir(model_dir_2)
+    assert model_exist_2==1, "model dir exist, sim time 86400"
+    ### time dir 2
+    t_dir_2=os.path.join(model_dir_2, 't000')
+    t_exist_2=os.path.isdir(t_dir_2)
[…]
