### 2026-02-04 e471c07
ball: removed norm by vol comment
