### 2026-02-04 229bc02
sld_grow: removed nom by vol of Iq
