### 2026-02-04 e15d94c
matplotlib in requirements.txt: reviewer comment implement
