### 2026-02-04 a599516
py pkgs op 2 remove, final check up done
