### 2026-02-04 #9772 @majumdararn
Dear reviewers,

I have addressed your concerns and made a patch release v1.0.1.

**[Issue 3](https://github.com/hereon-GEMS/Cont2SAS/issues/3#issue-3870170999) (Paper.md):**

See [comments](https://github.com/openjournals/joss-reviews/issues/9772#issuecomment-3817318552) above.

**[Issue 4](https://github.com/hereon-GEMS/Cont2SAS/issues/4#issue-3870214630) (Test scripts):**

1. Removed plot comparison tests by commenting.
2. Added reference data in tests/gold.
3. Added a test to compare the reference data with the generated data.

**[Issue 5](https://github.com/hereon-GEMS/Cont2SAS/issues/5#issue-3870227564) (Installation):**

Task 1 (Sassena appimage):

1. Mention OS req for appimage: Edited README mentioning Sassena.AppImage works only for Ubuntu 24.04.
2. Other OS: Added option 2 for other Linux-based OS, which involves precise instructions for cloning Sassena, building locally, copying the executable to the default appimage location and renaming it as appimage.

Task 2 (requirements.txt):

1. Changed the matplotlib version to 3.10.8, as suggested.
2. Removed the option of installing using pip inside conda.
