### 2026-02-13 #9772 @kremeyer
Hi @majumdararn,

I appreciate the quick turnaround on the raised issues regarding the Figures, tests and installation instructions.
Regarding the manuscript: The overall motivation is clear and the scope fits well for a short software paper. Most of my comments regard clarifications, tightening Figure-text consistency and notation, which I think could help readers outside the field appreciate the purpose of the software more. The numbers below refer to line numbers in the draft:

**Main comments**
23. You mention the "phase problem". I think a sentence, or just a clause, could make the connection of the phase loss in measuring intensities clear.
28. Define what "bigger features" refers to and explain why atomistic simulations are limited; when do molecular dynamics simulations become impractical?
48. You write that the user must provide SLD or “a set of variables from which SLD values can be calculated”. What are sets of variables that the SLD can be computed from? Is the computation of SLD implemented in Cont2SAS?
50. The text mentions time steps that are shown in Figure 1. It is not clear to me where those are.
50. Further, the right panel of Figure 1 contains the result of the main feature of the software, an I(Q) SAS signal. The authors should dedicate some attention to discussing this result and interpret, for example, the two minima in the plot.
53. The “chemical composition changes over time” is mentioned here. To me this felt like a bit out-of-context. I think an example would be helpful to motivate it.

**Minor editorial comments:**
Abstract: The abbreviations Cont, $I$, and $Q$ are introduced and never used.
22. Abbreviation TR is used and never used
47. $\beta$ is introduced and never used
Figure 2 caption. $\sigma_\mathrm{eff}$ is already introduced in the text
typos: "conituum", "can not", "taylor-made"; I suggest to carefully check for spelling, grammar and hyphenation again.

Regarding the repository itself I also have some minor comments:
- JOSS requires the documentation to also have the Statement of Need in it.
- In `CONTRIBUTING.md` you advise users to clone the repository and commit to it. That will not work without forking it first. I think it is not necessary to give a step-by-step explanation as these are standard steps.
- The automated tests all pass on my machine now. Since you are calling the scripts as a subprocess, pytest could not tell me anything about the coverage. From manually checking it looked like all the code is being run by the tests, great!
- The JOSS criteria expect an API documentation. Given that the code base has a manageable size and the functionality of all the modules is explained, the code accessible. Good comments in the code itself help new users. I consider this sufficient.

Once the above comments are addressed appropriately, I would go forward to recommending the paper to be accepted.
