+++ joss_submission/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 Small Angle Scattering (SAS) experiments with neutrons (SANS) and X-rays (SAXS) are useful techniques for probing material nanostructures [@chen2012characterizing]. With the Time Resolved (TR) variant of SAS, the time evolution of nanostructures can also be probed [@hollamby2013practical]. However, a direct retrieval of nanostructure from SAS data is not possible due to the so-called phase problem, i.e. the phase loss in measured intensities [@billinge2007problem]. Therefore, it is fruitful to combine SAS data with simulations to study nanostructures [@majumdar2024computation]. This approach is also useful for validating theories underlying simulations [@reich2022comparison].
 
-Simulated structures in real space can be generated using atomistic and continuum simulation. However, continuum simulation is the only option when the nanostructure evolution is influenced by bigger features or the SAS data is recorded for long time period, as atomistic simulation can not simulate large structures for long time [@ghavanloo2019computational]. Despite this indispensability, there is no software available for calculating SAS data from continuum simulations. `Cont2SAS` aims to fill this void by providing a tool that offers fast and error-free comparison of continuum simulations and SAS data. The ultimate goal is to study nanomaterials and validating theories using SAS data, particularly when length-scales and time-scales, inaccessible to atomistic simulations, need to be simulated.
+Simulated real-space structures can be generated using atomistic or continuum simulation. However, atomistic simulation cannot simulate large structures (e.g. millimeter scale) over long times (e.g. seconds) due to computational limitations [@ghavanloo2019computational]. This becomes problematic when the nanostructure evolution is influenced by larger length-scale structures or when the SAS data is recorded over long time periods. In such cases, continuum simulation is the only option. Despite this indispensability, there is no software available for calculating SAS data from continuum simulations. `Cont2SAS` aims to fill this void by providing a tool that offers fast and error-free comparison of continuum simulations and SAS data. The ultimate goal is to study nanomaterials and validating theories using SAS data, particularly when length-scales and time-scales, inaccessible to atomistic simulations, need to be simulated.
 
 # State of the field
 
