### 2026-02-15 eb4c079
comment 3: SLD calculation clarified
