+++ joss_submission/paper.md
@@ -45,7 +45,7 @@ From atomistic simulations, different software such as `nMoldyn` [@rog2003nmoldy
 
 ![Workflow of SAS pattern calculation: [left] SLD assignment for a spherical nanoparticle on generated mesh and [right] numerical calculation of SAS pattern. The numerical calculation is validated against known analytical formula for spherical nanoparticles [@guinier1955small]. \label{fig:sas_workflow}](figures/workflow.png)
 
-`Cont2SAS` calculates SAS patterns taking simulated nanostructures as an input. The simulated structure must provide either Scattering Length Density (SLD) ($\beta$) values or a set of variables from which SLD values can be calculated. The simulated input is processed to a data taylor-made for `Sassena`. `Sassena` calculates the SAS intensity $(I)$ as a function of scattering vector magnitude $(Q)$, i.e. SAS pattern, for different time steps (see \autoref{fig:sas_workflow}). 
+`Cont2SAS` calculates SAS patterns taking simulated nanostructures as an input. The simulated structure must include a Scattering Length Density (SLD) distribution, which is either directly obtained from simulation or calculated from a set of simulated variables (e.g. molar density and molar fraction) before being provided as input to `Cont2SAS`. The simulated input is processed to a data taylor-made for `Sassena`. `Sassena` calculates the SAS intensity $(I)$ as a function of scattering vector magnitude $(Q)$, i.e. SAS pattern, for different time steps (see \autoref{fig:sas_workflow}). 
 
 ![Workflow of effective cross-section $(\sigma_{\text{eff}})$ calculation: [left] SAS patterns at different timesteps ranging from 0 to 10 seconds and [right] $\sigma_{\text{eff}}$ calculated from SAS patterns. With time, the SLD of a spherical nanoparticle increases linearly. The calculated $\sigma_{\text{eff}}$ values are proportional to the ($\Delta$SLD)$^2$, i.e. square of SLD difference between the particle and its environment. \label{fig:sig_eff_workflow}](figures/sig_eff.png)
 
