+++ joss_submission/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 Small Angle Scattering (SAS) experiments with neutrons (SANS) and X-rays (SAXS) are useful techniques for probing material nanostructures [@chen2012characterizing]. With the Time Resolved (TR) variant of SAS, the time evolution of nanostructures can also be probed [@hollamby2013practical]. However, a direct retrieval of nanostructure from SAS data is not possible due to the so-called phase problem, i.e. the phase loss in measured intensities [@billinge2007problem]. Therefore, it is fruitful to combine SAS data with simulations to study nanostructures [@majumdar2024computation]. This approach is also useful for validating theories underlying simulations [@reich2022comparison].
 
-Simulated real-space structures can be generated using atomistic or continuum simulation. However, atomistic simulation cannot simulate large structures (e.g. millimeter scale) over long times (e.g. seconds) due to computational limitations [@ghavanloo2019computational]. This becomes problematic when the nanostructure evolution is influenced by larger length-scale structures or when the SAS data is recorded over long time periods. In such cases, continuum simulation is the only option. Despite this indispensability, there is no software available for calculating SAS data from continuum simulations. `Cont2SAS` aims to fill this void by providing a tool that offers fast and error-free comparison of continuum simulations and SAS data. The ultimate goal is to study nanomaterials and validating theories using SAS data, particularly when length-scales and time-scales, inaccessible to atomistic simulations, need to be simulated.
+Simulated real-space structures can be generated using atomistic or continuum simulation. However, atomistic simulation cannot feasibly simulate large structures (e.g. micrometer scale) over long times (e.g. seconds) due to computational limitations [@ghavanloo2019computational]. This becomes problematic when the nanostructure evolution is influenced by larger length-scale structures or when the SAS data is recorded over long time periods. In such cases, continuum simulation is the only option. Despite this indispensability, there is no software available for calculating SAS data from continuum simulations. `Cont2SAS` aims to fill this void by providing a tool that offers fast and error-free comparison of continuum simulations and SAS data. The ultimate goal is to study nanomaterials and validating theories using SAS data, particularly when length-scales and time-scales, inaccessible to atomistic simulations, need to be simulated.
 
 # State of the field
 
@@ -45,7 +45,7 @@ From atomistic simulations, different software such as `nMoldyn` [@rog2003nmoldy
 
 ![Workflow of SAS pattern calculation: [left] SLD assignment for a spherical nanoparticle on generated mesh and [right] numerical calculation of SAS pattern. The numerical calculation is validated against known analytical formula for spherical nanoparticles [@guinier1955small]. \label{fig:sas_workflow}](figures/workflow.png)
 
-`Cont2SAS` calculates SAS patterns taking simulated nanostructures as an input. The simulated structure must include a Scattering Length Density (SLD) distribution, which is either directly obtained from simulation or calculated from a set of simulated variables (e.g. molar density and molar fraction) before being provided as input to `Cont2SAS`. Simulating SLD distribution within `Cont2SAS` is also possible but only for model structures.  The simulated input is processed to a data tailor-made for `Sassena`. `Sassena` calculates SAS pattern, i.e. $I$ vs $Q$ data, as shown in \autoref{fig:sas_workflow}. The $Q$-dependence of $I$ reveals the average shape and size of the nanoparticles, whereas the magnitude of $I$ is related to the scattering contrast, i.e. the average diffrence in SLD between particles and their surrounding medium.
[…]
