### 2026-03-02 7b7e5a3
checking with sys.executable
