+++ CONTRIBUTING.md
@@ -107,4 +107,4 @@ Please follow the [Conventional Commits](https://www.conventionalcommits.org/) s
 
 <!-- omit in toc  -->
 ## Attribution
-This guide was generated by [contributing.md](https://contributing.md/generator). The authors performed necessary edits, and take full responsibility for the the content.
+This guide was generated by [contributing.md](https://contributing.md/generator). The authors performed necessary edits and take full responsibility for the the content.
