### 2026-03-03 38f035e
fixed ci error of conda
