+++ .github/workflows/ci.yml
@@ -60,7 +60,7 @@ jobs:
           activate-environment: Cont2Sas
           environment-file: environment.yml
           use-mamba: true
-          cache-environment: true
+          cache-environment: false
 
       # -----------------------
       # Tests
