+++ .github/workflows/ci.yml
@@ -1,6 +1,6 @@
 name: CI
 
-# Run on all branches, pushes, and PRs
+# Run on all branches and PRs
 on:
   push:
   pull_request:
@@ -10,71 +10,72 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  test:
-    name: Build & Test
+  download_sassena:
+    name: Download Sassena
     runs-on: ubuntu-latest
-    timeout-minutes: 30
-
-    env:
-      APPIMAGE_EXTRACT_AND_RUN: 1
-      C2S_HOME: ${{ github.workspace }}
-
     steps:
-      # -----------------------
       # Checkout code
-      # -----------------------
       - name: Checkout code
         uses: actions/checkout@v4
 
-      # -----------------------
-      # System dependencies
-      # -----------------------
-      - name: Install system packages
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y \
-            curl \
-            wget \
-            ghostscript \
-            mpi-default-dev
+      # Install curl
+      - name: Install curl
+        run: sudo apt-get update && sudo apt-get install -y curl
 
-      # -----------------------
       # Download Sassena AppImage
-      # -----------------------
       - name: Download Sassena AppImage
         run: |
           mkdir -p Sassena
-          curl -L \
-            https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage \
+          curl -L https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage \
             -o Sassena/Sassena.AppImage
           chmod +x Sassena/Sassena.AppImage
 
-      # -----------------------
-      # Miniforge + Conda environment
-      # -----------------------
-      - name: Setup Miniforge
-        uses: conda-incubator/setup-miniconda@v3
+      # Save as artifact for test job
+      - name: Upload Sassena artifact
+        uses: actions/upload-artifact@v3
         with:
-          miniforge-variant: Miniforge3
-          miniforge-version: latest
-          auto-activate-base: false
-          activate-environment: Cont2Sas
-          environment-file: environment.yml
-          use-mamba: true
-          cache-environment: false   # disable cache to prevent drift
+          name: Sassena
+          path: Sassena
 
-      # -----------------------
-      # Force requests install (optional)
-      # -----------------------
-      - name: Ensure requests is installed
+  run_tests:
+    name: Run Tests
+    runs-on: ubuntu-latest
+    needs: download_sassena
+    env:
+      APPIMAGE_EXTRACT_AND_RUN: 1
+      C2S_HOME: ${{ github.workspace }}
+    steps:
+      # Checkout code
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      # Install system dependencies
+      - name: Install system packages
+        run: sudo apt-get update && sudo apt-get install -y wget ghostscript mpi-default-dev
+
+      # Download Sassena artifact from previous job
+      - name: Download Sassena artifact
+        uses: actions/download-artifact@v3
+        with:
+          name: Sassena
+          path: Sassena
+
+      # Install Miniforge
+      - name: Install Miniforge
+        run: |
+          wget -O Miniforge3.sh "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
+          bash Miniforge3.sh -b -p "${HOME}/conda"
+          source "${HOME}/conda/etc/profile.d/conda.sh"
+
+      # Create and activate Conda environment
+      - name: Setup Conda environment
         shell: bash -l {0}
         run: |
[…]
