+++ .github/workflows/ci.yml
@@ -32,7 +32,7 @@ jobs:
 
       # Save as artifact for test job
       - name: Upload Sassena artifact
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: Sassena
           path: Sassena
@@ -55,7 +55,7 @@ jobs:
 
       # Download Sassena artifact from previous job
       - name: Download Sassena artifact
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           name: Sassena
           path: Sassena
