+++ .github/workflows/ci.yml
@@ -66,11 +66,6 @@ jobs:
           wget -O Miniforge3.sh "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
           bash Miniforge3.sh -b -p "${HOME}/conda"
           source "${HOME}/conda/etc/profile.d/conda.sh"
-
-      # Create and activate Conda environment
-      - name: Setup Conda environment
-        shell: bash -l {0}
-        run: |
           conda env create -f environment.yml
           conda activate Cont2Sas
           conda install -y requests>=2.31,<3
