+++ environment.yml
@@ -10,7 +10,6 @@ dependencies:
   - mdtraj=1.9.9
   - netCDF4=1.7.1 # 1.7.2 # 1.7.1.post2
   - numpy=1.26.4 # 2.3.1 # 1.26.4
-  - openmpi=5.0.8
   - pdf2image=1.17.0
   - PyMuPDF=1.26.1
   - pytest=7.4.4
