+++ .github/workflows/ci.yml
@@ -1,8 +1,10 @@
 name: CI
 
-# Run on all branches and PRs
 on:
   push:
+    branches:
+      - main
+      - develop 
   pull_request:
 
 concurrency:
@@ -10,67 +12,61 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  download_sassena:
-    name: Download Sassena
+  test:
+    name: Build & Test
     runs-on: ubuntu-latest
-    steps:
-      # Checkout code
-      - name: Checkout code
-        uses: actions/checkout@v4
+    timeout-minutes: 30
 
-      # Install curl
-      - name: Install curl
-        run: sudo apt-get update && sudo apt-get install -y curl
-
-      # Download Sassena AppImage
-      - name: Download Sassena AppImage
-        run: |
-          mkdir -p Sassena
-          curl -L https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage \
-            -o Sassena/Sassena.AppImage
-          chmod +x Sassena/Sassena.AppImage
-
-      # Save as artifact for test job
-      - name: Upload Sassena artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: Sassena
-          path: Sassena
-
-  run_tests:
-    name: Run Tests
-    runs-on: ubuntu-latest
-    needs: download_sassena
     env:
       APPIMAGE_EXTRACT_AND_RUN: 1
       C2S_HOME: ${{ github.workspace }}
+
     steps:
-      # Checkout code
       - name: Checkout code
         uses: actions/checkout@v4
 
-      # Install system dependencies
+      # -----------------------
+      # System dependencies
+      # -----------------------
       - name: Install system packages
-        run: sudo apt-get update && sudo apt-get install -y wget ghostscript mpi-default-dev
-
-      # Download Sassena artifact from previous job
-      - name: Download Sassena artifact
-        uses: actions/download-artifact@v4
-        with:
-          name: Sassena
-          path: Sassena
-
-      # Install Miniforge
-      - name: Install Miniforge
         run: |
-          wget -O Miniforge3.sh "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
-          bash Miniforge3.sh -b -p "${HOME}/conda"
-          source "${HOME}/conda/etc/profile.d/conda.sh"
-          conda env create -f environment.yml
-          conda activate Cont2Sas
-          conda install -y requests>=2.31,<3
+          sudo apt-get update
+          sudo apt-get install -y \
+            curl \
+            wget \
+            ghostscript \
+            mpi-default-dev
 
-      # Run pytest
-      - name: Run pytest
+      # -----------------------
+      # Download Sassena
+      # -----------------------
+      - name: Download Sassena AppImage
+        run: |
+          mkdir -p Sassena
+          curl -L \
+            https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage \
+            -o Sassena/Sassena.AppImage
+          chmod +x Sassena/Sassena.AppImage
+
+      # -----------------------
+      # Miniforge + Conda cache
+      # -----------------------
+      - name: Setup Miniforge
+        uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniforge-variant: Miniforge3
+          miniforge-version: latest
+          auto-activate-base: false
+          activate-environment: Cont2Sas
+          environment-file: environment.yml
+          use-mamba: true
+          cache-environment: true
+
+      # -----------------------
+      # Tests
[…]
