### 2026-03-03 5bd8fcb
uncover sld assignment from ball
