+++ .github/workflows/ci.yml
@@ -44,27 +44,26 @@ jobs:
             https://codebase.helmholtz.cloud/api/v4/projects/6801/packages/generic/sassena/v1.9.2-f31e3882/Sassena_CPU.AppImage \
             -o Sassena/Sassena.AppImage
           chmod +x Sassena/Sassena.AppImage
-          ./Sassena/Sassena.AppImage --help
 
-      # # -----------------------
-      # # Miniforge + Conda cache
-      # # -----------------------
-      # - name: Setup Miniforge
-      #   uses: conda-incubator/setup-miniconda@v3
-      #   with:
-      #     miniforge-variant: Miniforge3
-      #     miniforge-version: latest
-      #     auto-activate-base: false
-      #     activate-environment: Cont2Sas
-      #     environment-file: environment.yml
-      #     use-mamba: true
-      #     cache-environment: true
+      # -----------------------
+      # Miniforge + Conda cache
+      # -----------------------
+      - name: Setup Miniforge
+        uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniforge-variant: Miniforge3
+          miniforge-version: latest
+          auto-activate-base: false
+          activate-environment: Cont2Sas
+          environment-file: environment.yml
+          use-mamba: true
+          cache-environment: true
 
-      # # -----------------------
-      # # Tests
-      # # -----------------------
-      # - name: Run tests
-      #   shell: bash -l {0}
-      #   run: |
-      #     conda install -y pytest
-      #     pytest tests/test_ball.py -v -rsx
+      # -----------------------
+      # Tests
+      # -----------------------
+      - name: Run tests
+        shell: bash -l {0}
+        run: |
+          conda install -y pytest
+          Sassena/Sassena.AppImage --help # pytest tests/test_ball.py -v -rsx
