+++ .github/workflows/ci.yml
@@ -66,4 +66,5 @@ jobs:
         shell: bash -l {0}
         run: |
           conda install -y pytest
-          Sassena/Sassena.AppImage --help # pytest tests/test_ball.py -v -rsx
+          Sassena/Sassena.AppImage --help || true
+          pytest tests/test_ball.py -v -rsx
