+++ models/ball/ball_gen.py
@@ -176,13 +176,13 @@ ttot_scatt= round(toc_scatt - tic_scatt,3)
 print(f'Time taken for calculating SAS pattern: {ttot_scatt} s')
 print('-------------------------------------------------')
 
-# # Time satistics
-# print('')
-# print('-------------------------------------------------')
-# print('Time statistics')
-# print('-------------------------------------------------')
-# print(f'Mesh generation :{ttot_sg} s')
-# print(f'SLD assign      :{ttot_sim} s')
-# print(f'SAS calculation :{ttot_scatt} s')
-# print(f'Total           :{ttot_sg + ttot_sim + ttot_scatt} s')
-# print('-------------------------------------------------')
+# Time satistics
+print('')
+print('-------------------------------------------------')
+print('Time statistics')
+print('-------------------------------------------------')
+print(f'Mesh generation :{ttot_sg} s')
+print(f'SLD assign      :{ttot_sim} s')
+print(f'SAS calculation :{ttot_scatt} s')
+print(f'Total           :{ttot_sg + ttot_sim + ttot_scatt} s')
+print('-------------------------------------------------')
