### 2026-03-03 52ced82
changed ci commands only pytest runs
