+++ .github/workflows/ci.yml
@@ -66,7 +66,7 @@ jobs:
         shell: bash -l {0}
         run: |
           conda install -y pytest=7.4.4 setuptools
-          Sassena/Sassena.AppImage --help || true
-          # pytest tests/test_ball.py -v -rsx || true
+          # Sassena/Sassena.AppImage --help || true
+          pytest tests/test_ball.py -v -rsx
           # python models/ball/ball_gen.py
-          python src/scatt_cal.py 
+          # python src/scatt_cal.py 
