### 2026-03-03 924a27e
only scatt cal in ci
