+++ environment.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - pypi
 dependencies:
-  - python=3.12.1
+  - python=3.11
   - h5py=3.10.0 # 3.14.0 # 3.10.0
   - imageio=2.33.1 # 2.37.0 # 2.33.1
   - matplotlib=3.8.2 # 3.10.3 # 3.8.2
