### 2026-03-03 f7abd36
add error track in sub proc in test bal gen
