### 2026-03-03 660fbb3
ci back to pytest only
