### 2026-03-03 7f04ba5
capture ball scatt cal error
