### 2026-03-03 9dc4715
error tracking in plottting
