### 2026-03-03 a467444
added std err in ball gen err capture mode
