### 2026-03-03 a8317c9
remove branch rules from ci
