+++ .github/workflows/ci.yml
@@ -2,9 +2,6 @@ name: CI
 
 on:
   push:
-    branches:
-      - main
-      - develop 
   pull_request:
 
 concurrency:
