### 2026-03-05 1604fdc
removed error tracking
