### 2026-03-05 8834e15
run scatt cal script not pytest
