### 2026-03-05 b4dc7db
script run install setuptool
