+++ tests/test_ball.py
@@ -1,3 +1,176 @@
+# # Copyright (C) 2025  Helmholtz-Zentrum-Hereon
+
+# # This program is free software; you can redistribute it and/or
+# # modify it under the terms of the GNU General Public License
+# # as published by the Free Software Foundation; either version 2
+# # of the License, or (at your option) any later version.
+
+# # This program is distributed in the hope that it will be useful,
+# # but WITHOUT ANY WARRANTY; without even the implied warranty of
+# # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# # GNU General Public License for more details.
+
+# # You should have received a copy of the GNU General Public License
+# # along with this program; if not, see
+# # <https://www.gnu.org/licenses/>.
+
+# """
+# test script for checking one time step model
+# or structure model
+
+# chosen model: ball
+
+# created by: Arnab Majumdar
+# Date      : 06.08.2025
+# """
+
+# import shutil
+# import os
+# import subprocess
+# import sys
+# # import math
+# # import numpy as np
+# import pytest
+# # from matplotlib.testing.compare import compare_images as comp_img
+# import h5py
+
+# # non-test functions
+# def read_Iq_h5(Iq_data_h5):
+#     """
+#     Function for 
+#     1. Reading Iq data (Iq_data_h5)
+#     """
+#     Iq_data=h5py.File(Iq_data_h5,'r')
+#     Iq=Iq_data['Iq'][:]
+#     q=Iq_data['Q'][:]
+#     Iq_data.close()
+#     return q, Iq
+
+# # test functions
+# def test_ball_gen():
+#     """
+#     Test script for $C2S_HOME/models/ball/ball_gen.py:
+#     1. Check whether ball_gen runs
+#     """
+#     #     try:
+#     #     result = subprocess.run(
+#     #         ["python", script_path],
+#     #         capture_output=True,   # capture stdout and stderr
+#     #         text=True,             # decode bytes to string
+#     #         check=True             # raise CalledProcessError on non-zero exit
+#     #     )
+#     #     print("STDOUT:", result.stdout)
+#     # except subprocess.CalledProcessError as e:
+#     #     print("Command failed!")
+#     #     print("Return code:", e.returncode)
+#     #     print("STDOUT:", e.stdout)
+#     #     print("STDERR:", e.stderr)
+#     try:
+#         gen_result = subprocess.run(
+#             [sys.executable, "models/ball/ball_gen.py"],
+#             capture_output=True,
+#             text=True,
+#             check=True
+#         )
+#         print("STDOUT:", gen_result.stdout)
+#     except subprocess.CalledProcessError as e:
+#         print("Command failed!")
+#         print("Return code:", e.returncode)
+#         print("STDOUT:", e.stdout)
+#         print("STDERR:", e.stderr)
+
+#     # Ensure it runs without crashing
+#     assert gen_result.returncode == 0, "ball model did not generate data"
+
+# def test_ball_plot():
+#     """
+#     Test script for $C2S_HOME/models/ball/ball_plot.py:
+#     1. Check whether ball_plot runs
+#     """
+#     try:
+#         plot_result = subprocess.run(
+#             ["python", "models/ball/ball_plot.py"],
+#             capture_output=True,
+#             text=True,
+#             check=True
+#         )
+#         print("STDOUT:", plot_result.stdout)
+#     except subprocess.CalledProcessError as e:
[…]
