+++ .github/workflows/ci.yml
@@ -67,6 +67,8 @@ jobs:
         run: |
           conda install -y setuptools
           # Sassena/Sassena.AppImage --help || true
-          pytest tests/test_ball.py -v -rsx
+          # pytest tests/test_ball.py -v -rsx
           # python models/ball/ball_gen.py
-          # python src/scatt_cal.py 
+          python src/struct_gen.py
+          python src/sim_gen.py
+          python src/scatt_cal.py 
