### 2026-03-05 2a751b6
added run complete msg in sim gen
