### 2026-03-05 56197c2
added debug msgs
