+++ .github/workflows/ci.yml
@@ -66,6 +66,8 @@ jobs:
         shell: bash -l {0}
         run: |
           conda install -y setuptools
+          which python
+          conda list setuptools
           # Sassena/Sassena.AppImage --help || true
           # pytest tests/test_ball.py -v -rsx
           # python models/ball/ball_gen.py
