### 2026-03-05 b90bb99
pytest for all models not only balls
