+++ .github/workflows/ci.yml
@@ -69,8 +69,8 @@ jobs:
           # which python
           # conda list setuptools
           # Sassena/Sassena.AppImage --help || true
-          pytest tests/test_ball.py -v -rsx
-          # python models/ball/ball_gen.py
+          pytest -v -rsx
+          # python models/ball/bally_gen.py
           # python -c "import mdtraj; print('mdtraj OK')"
           # python src/struct_gen.py
           # python src/sim_gen.py
