### 2026-03-06 9ddd39f
added openmpi
