### 2026-03-06 6fb5add
removed error tracking from ball gen
