### 2026-03-06 4ed9f00
added end message in sim gen script
