+++ src/scatt_cal.py
@@ -308,5 +308,3 @@ for i, t in enumerate(t_arr):
     Iq_data['Iq']=Iq
     Iq_data['Q']=q
     Iq_data.close()
-
-    print('simulation gen run complete')
