+++ tests/test_ball.py
@@ -52,19 +52,6 @@ def test_ball_gen():
     Test script for $C2S_HOME/models/ball/ball_gen.py:
     1. Check whether ball_gen runs
     """
-    #     try:
-    #     result = subprocess.run(
-    #         ["python", script_path],
-    #         capture_output=True,   # capture stdout and stderr
-    #         text=True,             # decode bytes to string
-    #         check=True             # raise CalledProcessError on non-zero exit
-    #     )
-    #     print("STDOUT:", result.stdout)
-    # except subprocess.CalledProcessError as e:
-    #     print("Command failed!")
-    #     print("Return code:", e.returncode)
-    #     print("STDOUT:", e.stdout)
-    #     print("STDERR:", e.stderr)
     try:
         gen_result = subprocess.run(
             [sys.executable, "models/ball/ball_gen.py"],
