### 2026-03-06 824f355
added error tracking in phase field
