### 2026-03-06 4954238
added error tracking in sld grow
