### 2026-03-09 #9772 @majumdararn
Dear reviewers,

Thank you for your comments so far.
I have implemented the necessary changes to address the comments from @kremeyer.
The changed version is tagged as `v1.0.2`.

The following is the summary:

1. I added a clause after mentioning 'phase problem'.
2. Reformulated sentence differently to explain 'bigger feature'. Also mentioned is at what length and time scale MD simulation becomes impractical.
3. I included an example variable set to demonstrate how SLD can be calculated from other variables. I mentioned clearly what is implemented in Cont2SAS and what needs to be done before giving input to Cont2SAS.
4. The mention of time steps is moved to figure 2 because figure 1 shows no time-evolution.
5. An analysis of the right-hand plot in figure 1 is added.
6. The chemical composition change is replaced with SLD, shape, and size description. Both are explaining the same thing differently. The current description is more relatable because SLD, shape, and size are mentioned before in the manuscript.
7. TR is used in the manuscript.
8. &beta; is deleted.
9. Double introduction of &sigma;<sub>eff</sub> is removed. It is now introduced in the text.
10. "conituum", "can not", "taylor-made" corrected.
11. Statement of need is added in documentation
12. Mentioned about forking in `CONTRIBUTING.md`
13. Added error tracking in tests using `try` and `except`

Note: I would like to thank @cchndl from Forschungszentrtum Jülich for his advice and support while migrating this project from GitLab to GitHub.
