### 2026-03-25 #9772 @kremeyer
With the latest revision to `v1.0.2` all of my main comments were addressed to a statisfactory degree. In my last top-to-bottom read however, I still found a handful of typos and grammar mistakes that need to be addressed, especially around missing or extra articles like:
- `generating a TR-SAS data` should be either `generating a TR-SAS dataset` or `generating TR-SAS data`
- `diffrence` should be `difference`

That leaves my checklist with only the Quality of Writing point unchecked.
