### 2026-06-11 2a48c5c
added references
