+++ joss_submission/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 # Summary
 
-`Cont2SAS` is a software tool built around the existing software solution `Sassena` -- known for calculating scattering patterns from simulated atomic structures. The goal of `Cont2SAS` is to provide a software similar to `Sassena`, but for calculating Small Angle Scattering data from simulated continuum nanostructures. `Cont2SAS` can calculate SAS patterns, i.e. SAS intensity $(I)$ vs. scattering vector magnitude $(Q)$ using a novel numerical method. `Cont2SAS` can also calculate effective cross-section $(\sigma_{\text{eff}})$, i.e. the count rate of scattered radiation per incident unit flux. $\sigma_{\text{eff}}$ is calculated by integrating instrument-agnostic SAS patterns taking instrument geometry into account. `Cont2SAS` can be used for different purposes, such as validating simulations, tuning simulation parameters, and analyzing SAS data.
+`Cont2SAS` is software tool built around the existing software solution `Sassena` -- known for calculating scattering patterns from simulated atomic structures. The goal of `Cont2SAS` is to provide software similar to `Sassena`, but for calculating Small Angle Scattering data from simulated continuum nanostructures. `Cont2SAS` can calculate SAS patterns, i.e. SAS intensity $(I)$ vs. scattering vector magnitude $(Q)$ using a novel numerical method. `Cont2SAS` can also calculate effective cross-section $(\sigma_{\text{eff}})$, i.e. the count rate of scattered radiation per incident unit flux. $\sigma_{\text{eff}}$ is calculated by integrating instrument-agnostic SAS patterns taking instrument geometry into account. `Cont2SAS` can be used for different purposes, such as validating simulations, tuning simulation parameters, and analyzing SAS data.
 
 # Statement of need
 
@@ -40,21 +40,21 @@ Simulated real-space structures can be generated using atomistic or continuum si
 
 # State of the field
 
-From atomistic simulations, different software such as `nMoldyn` [@rog2003nmoldyn; @nmoldyn2023hinsen], `MDANSE` [@goret2017mdanse; @mdanse2023isis], `LiquidLib` [@walter2018liquidlib; @liquidlib2023zlab], and `Sassena` [@lindner2012sassena; @lindner2012towards; @sassena2017lindner; @sassena2023majumdar; @majumdar2024computation] can calculate SAS patterns but they cannot calculate the same from continuum simulations. An in-house code is available for continuum simulations but it was built for a 2D membrane system lacking generality [@dorrell2020combined]. Hence, `Cont2SAS` is built using `Sassena` as a backend calculator. `Sassena` was chosen over other software due to its notable computation speed and robustness [@majumdar2024computation].
+From atomistic simulations, different software such as `nMoldyn` [@rog2003nmoldyn; @nmoldyn2023hinsen], `MDANSE` [@goret2017mdanse; @mdanse2023isis], `LiquidLib` [@walter2018liquidlib; @liquidlib2023zlab], and `Sassena` [@lindner2012sassena; @lindner2012towards; @sassena2017lindner; @sassena2023majumdar; @majumdar2024computation] can calculate SAS patterns but they cannot calculate the same from continuum simulations. Some in-house codes are available for continuum simulations but they lack reusability [schmidt2007simulation; @dorrell2020combined; berbenni2025simulation]. Hence, `Cont2SAS` is built using `Sassena` as a backend calculator. `Sassena` was chosen over other software due to its notable computation speed and robustness [@majumdar2024computation].
 
 # Software features
 
[…]
