### 2026-06-19 a8635db
changed figure width after merging, this figure is added in paper
