+++ joss_submission/paper.bib
@@ -165,3 +165,24 @@ doi = {10.3233/JNR-190116}
   year={1955},
   publisher={Wiley New York}
 }
+@article{schmidt2007simulation,
+  title={Simulation of small-angle scattering curves by numerical Fourier transformation},
+  author={Schmidt-Rohr, Klaus},
+  journal={Journal of Applied Crystallography},
+  volume={40},
+  number={1},
+  pages={16--25},
+  year={2007},
+  publisher={International Union of Crystallography},
+  doi={10.1107/S002188980604550X}
+}
+@article{berbenni2025simulation,
+  title={Simulation and analysis of small angle scattering (SAS) patterns of Ni-based superalloy microstructures generated by a phase-field model},
+  author={Berbenni, St{\'e}phane and Dingreville, R{\'e}mi and Schenk, Thomas},
+  journal={Computational Materials Science},
+  volume={258},
+  pages={114086},
+  year={2025},
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2025.114086}
+}
