### 2026-06-19 c3d018a
minor adjustment in fi2 left subfig
