### 2026-06-19 338fbd0
readjusted legend for fig2 left subfig
