### 2026-06-19 9c2cde2
fig 2 caption: mentioned proportionality constant
