+++ joss_submission/paper.md
@@ -48,7 +48,7 @@ From atomistic simulations, different software such as `nMoldyn` [@rog2003nmoldy
 
 `Cont2SAS` calculates SAS patterns taking simulated nanostructures as an input. The simulated structure must include a Scattering Length Density (SLD) distribution, which is either directly obtained from simulation or calculated from a set of simulated variables (e.g. molar density and molar fraction) before being provided as input to `Cont2SAS`. Simulating SLD distribution within `Cont2SAS` is also possible for model structures.  The simulated input is processed to data tailor-made for `Sassena`. `Sassena` calculates SAS pattern, i.e. $I$ vs $Q$ data, as shown in \autoref{fig:sas_workflow}. The $Q$-dependence of $I$ reveals the average shape and size of the nanoparticles, whereas the magnitude of $I$ is related to the scattering contrast, i.e. the average difference in SLD between particles and their surrounding medium.
 
-![[left] TR-SAS data, i.e. SAS patterns at different timesteps ranging from 0 to 10 seconds. With time, the SLD of a spherical nanoparticle increases linearly. [right] $\sigma_{\text{eff}}$ calculated from SAS patterns. The calculated $\sigma_{\text{eff}}$ values are proportional to the ($\Delta$SLD)$^2$, i.e. square of SLD difference between the particle and its environment. \label{fig:sig_eff_workflow}](figures/sig_eff.png)
+![[left] TR-SAS data, i.e. SAS patterns at different timesteps ranging from 0 to 10 seconds. With time, the SLD of a spherical nanoparticle increases linearly. [right] $\sigma_{\text{eff}}$ calculated from SAS patterns. The calculated $\sigma_{\text{eff}}$ values are proportional to the ($\Delta$SLD)$^2$, i.e. square of SLD difference between the particle and its environment. The shown ($\Delta$SLD)$^2$ values are multiplied by a propotionality constant. \label{fig:sig_eff_workflow}](figures/sig_eff.png)
 
 For a simulation with more than one timestep, SAS patterns can be calculated from different time steps generating a TR-SAS dataset, as shown in \autoref{fig:sig_eff_workflow}. From TR-SAS data, `Cont2SAS` can calculate the time evolution of effective cross-section $(\sigma_{\text{eff}})$, which is defined as the count rate per incident unit flux (see \autoref{fig:sig_eff_workflow}). This feature was not available in pervious in-house implementations [@schmidt2007simulation; @dorrell2020combined; @berbenni2025simulation] or `Sassena` [@sassena2023majumdar]. The time evolution of count rate is useful when the SLD of nanostructure changes over time without any change in average shape and size, e.g. while storing hydrogen in ball-milled powder sample [@aslan2019high]. The calculated $\sigma_{\text{eff}}$ must be multiplied by an empirical factor before comparing with measured count rate.   
 
