### 2026-06-19 ae317cc
fix fig 2 caption to add prop cons
