### 2026-06-19 09bfaab
changed fig2 axes ticks
