### 2026-06-22 2c3af83
cracked netcdf4 up in conda installation
