+++ environment.yml
@@ -8,7 +8,7 @@ dependencies:
   - imageio=2.33.1 # 2.37.0 # 2.33.1
   - matplotlib=3.8.2 # 3.10.3 # 3.8.2
   - mdtraj=1.9.9
-  - netCDF4=1.7.1 # 1.7.2 # 1.7.1.post2
+  - netCDF4=1.7.4 #1.7.1 # 1.7.2 # 1.7.1.post2
   - numpy=1.26.4 # 2.3.1 # 1.26.4
   - openmpi=5.0.8
   - pdf2image=1.17.0
