+++ environment.yml
@@ -3,17 +3,17 @@ channels:
   - conda-forge
   - pypi
 dependencies:
-  - python=3.12.1
-  - h5py=3.10.0 # 3.14.0 # 3.10.0
-  - imageio=2.33.1 # 2.37.0 # 2.33.1
-  - matplotlib=3.8.2 # 3.10.3 # 3.8.2
-  - mdtraj=1.9.9
-  - netCDF4=1.7.4 #1.7.1 # 1.7.2 # 1.7.1.post2
-  - numpy=1.26.4 # 2.3.1 # 1.26.4
-  - openmpi=5.0.8
-  - pdf2image=1.17.0
-  - PyMuPDF=1.26.1
-  - pytest=7.4.4
-  - pylint=3.3.8
-  - scipy=1.11.4 # 1.16.0 # 1.11.4
-  - pip=23.3.2
+  - python #=3.12 #3.12.1
+  - h5py #=3.10.0 # 3.14.0 # 3.10.0
+  - imageio #=2.33.1 # 2.37.0 # 2.33.1
+  - matplotlib #=3.8.2 # 3.10.3 # 3.8.2
+  - mdtraj #=1.9.9
+  - netCDF4 #=1.7.4 #1.7.1 # 1.7.2 # 1.7.1.post2
+  - numpy #=1.26.4 # 2.3.1 # 1.26.4
+  - openmpi #=5.0.8
+  - pdf2image #=1.17.0
+  - PyMuPDF #=1.26.1
+  - pytest #=7.4.4
+  - pylint #=3.3.8
+  - scipy #=1.12 #1.11.4 # 1.16.0 # 1.11.4
+  - pip #=23.3.2
