### 2026-06-22 fcd1bfa
changed get_cmap to new colormaps style
