### 2026-06-22 374c077
setuptools installation moved to env.yml
