### 2026-06-22 0b853bc
added setuptools for conda and rearrange
