+++ environment.yml
@@ -12,8 +12,11 @@ dependencies:
   - numpy #=1.26.4 # 2.3.1 # 1.26.4
   - openmpi #=5.0.8
   - pdf2image #=1.17.0
+  - pip #=23.3.2
   - PyMuPDF #=1.26.1
   - pytest #=7.4.4
   - pylint #=3.3.8
+  - setuptools<=80.10.2
   - scipy #=1.12 #1.11.4 # 1.16.0 # 1.11.4
-  - pip #=23.3.2
+  
+  
