### 2026-06-23 609b80e
fix sassena installation: issue 16
