### 2026-06-24 1397d7e
changed figure2 fonts again
