### 2026-06-24 b9d32b4
added minor changes after auth3 suggestion
