+++ joss_submission/paper.bib
@@ -103,12 +103,6 @@ doi = {10.3233/JNR-190116}
   howpublished   = "\url{https://codebase.helmholtz.cloud/DAPHNE4NFDI/sassena}",
   doi={10.5281/zenodo.10037485}
 }
-@misc{sassena2017lindner,
-  author = {Lindner, Benjamin},
-  title = {Sassena},
-  year  = {2017},
-  howpublished   = "\url{https://github.com/camm/sassena}",
-}
 @article{rog2003nmoldyn,
   title={nMoldyn: a program package for a neutron scattering oriented analysis of molecular dynamics simulations},
   author={R{\'o}g, Tomasz and Murzyn, Krzysztof and Hinsen, Konrad and Kneller, Gerald R},
@@ -138,8 +132,8 @@ doi = {10.3233/JNR-190116}
   doi={10.1021/acs.jcim.6b00571}
 }
 @misc{mdanse2023isis,
-  author = {UKRI STFC ISIS Neutron and Muon Facility},
-  title = {MDANSE--Version 1.5.2},
+  author = {{UKRI STFC ISIS Neutron and Muon Facility}},
+  title = {MDANSE--Version 2.0.1},
   year  = {2023},
   howpublished   = "\url{https://github.com/ISISNeutronMuon/MDANSE/}",
 }
