+++ joss_submission/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Cont2SAS: A python package for calculating SAS parameters from continuum nanostructures'
+title: 'Cont2SAS: A Python package for calculating small angle scattering parameters from continuum nanostructures'
 tags:
   - Python
   - Continuum simulation
@@ -30,31 +30,31 @@ bibliography: paper.bib
 
 # Summary
 
-`Cont2SAS` is a software tool built around the existing software solution `Sassena` -- known for calculating scattering patterns from simulated atomic structures. The goal of `Cont2SAS` is to provide software similar to `Sassena`, but for calculating Small Angle Scattering (SAS) data from simulated continuum nanostructures. `Cont2SAS` can calculate SAS patterns, i.e. SAS intensity $(I)$ vs. scattering vector magnitude $(Q)$ using a novel numerical method. `Cont2SAS` can also calculate effective cross-section $(\sigma_{\text{eff}})$, i.e. the count rate of scattered radiation per incident unit flux. $\sigma_{\text{eff}}$ is calculated by integrating instrument-agnostic SAS patterns taking instrument geometry into account. `Cont2SAS` can be used for different purposes, such as validating simulations, tuning simulation parameters, and analyzing SAS data.
+`Cont2SAS` is a software tool built around the existing software solution `Sassena` -- known for calculating scattering patterns from simulated atomic structures. The goal of `Cont2SAS` is to provide software similar to `Sassena`, but for calculating Small Angle Scattering (SAS) data from simulated continuum nanostructures. `Cont2SAS` can calculate SAS patterns, i.e. SAS intensity $(I)$ vs. scattering vector magnitude $(Q)$ using numerical method. `Cont2SAS` can also calculate effective cross-section $(\sigma_{\text{eff}})$, i.e. the count rate of scattered radiation per incident unit flux. $\sigma_{\text{eff}}$ is calculated by integrating instrument-agnostic SAS patterns taking the instrument geometry into account. `Cont2SAS` can be used for different purposes, such as validating simulations, tuning simulation parameters, and analyzing SAS data.
 
 # Statement of need
 
 SAS experiments with neutrons (SANS) and X-rays (SAXS) are useful techniques for probing material nanostructures [@chen2012characterizing]. With the Time Resolved (TR) variant of SAS, the time evolution of nanostructures can also be probed [@hollamby2013practical]. However, a direct retrieval of nanostructure from SAS data is not possible due to the so-called phase problem, i.e. the phase loss in measured intensities [@billinge2007problem]. Therefore, it is fruitful to combine SAS data with simulations to study nanostructures [@majumdar2024computation]. This approach is also useful for validating theories underlying simulations [@reich2022comparison].
 
-Simulated real-space structures can be generated using atomistic or continuum simulations. However, atomistic simulations cannot feasibly simulate large structures (e.g. micrometer scale) over long times (e.g. seconds) due to computational limitations [@ghavanloo2019computational]. This becomes problematic when nanostructure evolution is influenced by larger length-scale structures or when the SAS data is recorded over long time periods. In such cases, continuum simulation is the only option. Despite this indispensability, there is no software available for calculating SAS data from continuum simulations. `Cont2SAS` aims to fill this void by providing a tool that offers fast and error-free comparison of continuum simulations and SAS data. The ultimate goal is to study nanomaterials and validate theories using SAS data, particularly when length-scales and time-scales, inaccessible to atomistic simulations, need to be simulated.
[…]
