+++ joss_submission/paper.bib
@@ -180,3 +180,24 @@ doi = {10.3233/JNR-190116}
   publisher={Elsevier},
   doi={10.1016/j.commatsci.2025.114086}
 }
+@misc{liu2023generic,
+  author = {Liu, Yun and Bourne, Robert and King, Steve},
+  title = {Generic SAS Calculator Tool},
+  year  = {2023},
+  howpublished   = "\url{https://www.sasview.org/docs/user/qtgui/Calculators/sas_calculator_help.html}",
+}
+@misc{sasview,
+  author = {{SasView Developers}},
+  title = {SasView--Version 6.1.3},
+  year  = {2026},
+  howpublished   = "\url{https://www.sasview.org}",
+}
+@misc{majumdar2026characterization,
+  title        = {Characterization of nanomaterials through complementary use of continuum simulations and small angle scattering measurements},
+  author       = {Majumdar, Arnab and Müller, Martin and Busch, Sebastian},
+  year         = {2026},
+  howpublished = {SSRN working paper},
+  url          = {https://ssrn.com/abstract=6460568},
+  doi          = {10.2139/ssrn.6460568},
+  note         = {Available at SSRN}
+}
