### 2026-06-26 32bfa70
added auth3 comments related to sasview
