+++ joss_submission/paper.md
@@ -40,7 +40,7 @@ Simulated real-space structures can be generated using atomistic or continuum si
 
 # State of the field
 
-From atomistic simulations, different software packages such as `nMoldyn` [@rog2003nmoldyn; @nmoldyn2023hinsen], `MDANSE` [@goret2017mdanse; @mdanse2023isis], `LiquidLib` [@walter2018liquidlib; @liquidlib2023zlab], and `Sassena` [@lindner2012sassena; @lindner2012towards; @sassena2023majumdar; @majumdar2024computation] can calculate SAS patterns but they cannot calculate the same from continuum simulations. For continuum simulations, existing software implementations are typically developed for problem-specific applications [@schmidt2007simulation; @dorrell2020combined; @berbenni2025simulation]. The `Generic SAS Calculator` option in `SasView` [@sasview] is also a potential alternative, but lacks a clear strategy for converting simulation output into the expected input formats, `grid type data` and `element type data` [@liu2023generic]. Hence, `Cont2SAS` is built as a general-purpose program that converts simulation output into discretized atoms using a well-defined strategy [@majumdar2026characterization]. For calculating SAS data from the discretized atoms, `Cont2SAS` uses `Sassena` as a backend calculator, which provides notable computation speed and robustness [@majumdar2024computation].
+Different software packages such as `nMoldyn` [@rog2003nmoldyn; @nmoldyn2023hinsen], `MDANSE` [@goret2017mdanse; @mdanse2023isis], `LiquidLib` [@walter2018liquidlib; @liquidlib2023zlab], and `Sassena` [@lindner2012sassena; @lindner2012towards; @sassena2023majumdar; @majumdar2024computation] can calculate SAS patterns from atomistic simulations, but they cannot calculate the same from continuum simulations. For continuum simulations, existing software implementations are typically developed for problem-specific applications [@schmidt2007simulation; @dorrell2020combined; @berbenni2025simulation]. The `Generic SAS Calculator` option in `SasView` [@sasview] is also a potential alternative, but lacks a clear strategy for converting simulation output into the supported input formats [@liu2023generic]. Hence, `Cont2SAS` is built as a general-purpose program that converts simulation output into discretized atoms using a well-defined strategy [@majumdar2026characterization]. For calculating SAS data from the discretized atoms, `Cont2SAS` uses `Sassena` as a backend calculator, which provides notable computation speed and robustness [@majumdar2024computation].
 
 # Software features
 
