### 2026-06-26 e3d58a3
cleaned ci script
