+++ .github/workflows/ci.yml
@@ -65,13 +65,4 @@ jobs:
       - name: Run tests
         shell: bash -l {0}
         run: |
-          # conda install -y "setuptools<81"
-          # which python
-          # conda list setuptools
-          # Sassena/Sassena.AppImage --help || true
-          pytest -v -rsx
-          # python models/ball/bally_gen.py
-          # python -c "import mdtraj; print('mdtraj OK')"
-          # python src/struct_gen.py
-          # python src/sim_gen.py
-          # python src/scatt_cal.py 
+          pytest -v -rsx 
