+++ joss_submission/paper.md
@@ -40,13 +40,13 @@ Simulated real-space structures can be generated using atomistic or continuum si
 
 # State of the field
 
-Different software packages such as `nMoldyn` [@rog2003nmoldyn; @nmoldyn2023hinsen], `MDANSE` [@goret2017mdanse; @mdanse2023isis], `LiquidLib` [@walter2018liquidlib; @liquidlib2023zlab], and `Sassena` [@lindner2012sassena; @lindner2012towards; @sassena2023majumdar; @majumdar2024computation] can calculate SAS patterns from atomistic simulations, but they cannot calculate the same from continuum simulations. For continuum simulations, existing software implementations are typically developed for problem-specific applications [@schmidt2007simulation; @dorrell2020combined; @berbenni2025simulation]. The `Generic SAS Calculator` option in `SasView` [@sasview] is also a potential alternative, but lacks a clear strategy for converting simulation output into the supported input formats [@liu2023generic]. Hence, `Cont2SAS` is built as a general-purpose program that converts simulation output into discretized atoms using a well-defined strategy [@majumdar2026characterization]. For calculating SAS data from the discretized atoms, `Cont2SAS` uses `Sassena` as a backend calculator, which provides notable computation speed and robustness [@majumdar2024computation].
+Different software packages such as `nMoldyn` [@rog2003nmoldyn; @nmoldyn2023hinsen], `MDANSE` [@goret2017mdanse; @mdanse2023isis], `LiquidLib` [@walter2018liquidlib; @liquidlib2023zlab], and `Sassena` [@lindner2012sassena; @lindner2012towards; @sassena2023majumdar; @majumdar2024computation] can calculate SAS patterns from atomistic simulations, but they cannot calculate the same from continuum simulations. For continuum simulations, existing software implementations are typically developed for problem-specific applications [@schmidt2007simulation; @dorrell2020combined; @berbenni2025simulation]. The `Generic SAS Calculator` option in `SasView` [@sasview] is also a potential alternative. However, `SasView` doesn't provide an option to convert simulation output into either discretized points or elements with constant Scattering Length Density (SLD) required by the `Generic SAS Calculator` [@liu2023generic]. Hence, `Cont2SAS` is built as a general-purpose program that converts simulation output into discretized atoms using a well-defined strategy [@majumdar2026characterization]. For calculating SAS data from the discretized atoms, `Cont2SAS` uses `Sassena` as a backend calculator, which provides notable computation speed and robustness [@majumdar2024computation].
 
 # Software features
 
 ![Workflow of calculating SAS patterns: [left] SLD assignment for a spherical nanoparticle on generated mesh and [right] numerical calculation of SAS pattern. The numerical calculation is validated against the known analytical formula for spherical nanoparticles [@guinier1955small]. \label{fig:sas_workflow}](figures/workflow.png)
 
-`Cont2SAS` calculates SAS patterns taking simulated nanostructures as input. The simulated structure must include a Scattering Length Density (SLD) distribution, which is either directly obtained from simulation or calculated from a set of simulated variables (e.g. molar density and molar fraction) before being provided as input to `Cont2SAS`. Simulating SLD distributions is also possible within `Cont2SAS` for model structures.  The simulated input is processed to data tailor-made for `Sassena`, which calculates SAS patterns, i.e. $I$ vs $Q$ data, as shown in \autoref{fig:sas_workflow}. The $Q$-dependence of $I$ reveals the average shape and size of the nanoparticles, whereas the magnitude of $I$ is related to the scattering contrast, i.e. the average difference in SLD between particles and their surrounding medium.
[…]
