### 2026-07-20 576c01d
added 2 more citations for nmoldyn
