### 2026-08-04 3c64996
changed lindner thesis doi to url
