+++ joss_submission/paper.bib
@@ -105,7 +105,7 @@
                 Aoun, Bachir and Murzyn, Krzysztof and R{\'o}g, Tomasz},
   title = {nMOLDYN3--Version 3.0.12},
   year  = {2023},
-  howpublished   = "\url{https://github.com/khinsen/nMOLDYN3/}",
+  howpublished   = "\url{https://github.com/khinsen/nMOLDYN3/}"
 }
 @article{goret2017mdanse,
   title={MDANSE: An interactive analysis environment for molecular dynamics simulations},
@@ -142,7 +142,7 @@
   author = {Z-laboratory},
   title = {Liquidlib--Version 1.0.0},
   year  = {2022},
-  howpublished   = "\url{https://z-laboratory.github.io/LiquidLib/}",
+  howpublished   = "\url{https://z-laboratory.github.io/LiquidLib/}"
 }
 @article{lindner2012sassena,
   title={Sassena—X-ray and neutron scattering calculated from molecular dynamics trajectories using massively parallel computers},
@@ -160,7 +160,7 @@
   author={Lindner, Benjamin},
   school={University of Tennessee, United States},
   year={2012},
-  doi={20.500.14382/22524}
+  howpublished="\url{https://trace.tennessee.edu/handle/20.500.14382/22524}"
 }
 @misc{sassena2023majumdar,
   author = {Majumdar, Arnab and Lindner, Benjamin},
@@ -210,7 +210,7 @@
   author = {Liu, Yun and Bourne, Robert and King, Steve},
   title = {Generic SAS Calculator Tool},
   year  = {2023},
-  howpublished   = "\url{https://www.sasview.org/docs/user/qtgui/Calculators/sas_calculator_help.html}",
+  howpublished   = "\url{https://www.sasview.org/docs/user/qtgui/Calculators/sas_calculator_help.html}"
 }
 @misc{majumdar2026characterization,
   title        = {Characterization of nanomaterials through complementary use of continuum simulations and small angle scattering measurements},
