### 2026-08-18 #9772 @rkurchin
Hi @majumdararn, I'll do some other checks shortly, but please check on the flagged DOI's [above](https://github.com/openjournals/joss-reviews/issues/9772#issuecomment-5270362121).
