### 2026-08-18 #9772 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] Archive listed title and authors matches paper --> **Please make the archive title match the paper title**
