### 2026-08-18 #9772 @majumdararn
Dear @rkurchin,

Thank you for the final check. Here are my comments:

**Comment 1**

> Archive listed title and authors matches paper --> Please make the archive title match the paper title

done

**Comment 2**

> Hi @majumdararn, I'll do some other checks shortly, but please check on the flagged DOI's https://github.com/openjournals/joss-reviews/issues/9772#issuecomment-5270362121.

The flagged DOI (https://doi.org/10.1016/B978-0-408-70077-1.50036-1) points to another book written by A. RENOUPREZ. This book is different from the book we cited, but it bears the same name.

We cited:

> Guinier, A., Fournet, G., Walker, C. B., & Yudowitch, K. L. (1955). Small-angle scattering of
> x-rays. Wiley New York.

We could not find a DOI for this book. This could be because this book is very old (1955). There are some review articles with DOIs, but the book itself does not have a DOI. This was also checked during the review and was discussed [here](https://github.com/openjournals/joss-reviews/issues/9772#issuecomment-4852012649).
