### 2026-08-25 #9772 @rkurchin
Congrats to @majumdararn and thanks to handling editor @jgostick and reviewers @ebknudsen and @kremeyer. JOSS wouldn't exist without your volunteer efforts!
